Integer division in Python 3
Under the _Floating Point Numbers_ section, the notebook wants to
illustrate that regular division of whole numbers would produce '0' if
the code is run, as opposed to producing '0.125' when users divide with
floating numbers. However, running the code in the In [ ] box printed
'0.125' for both lines.

Dividing with whole numbers yields floats in
the subsequent inputs in the notebook. This doesn't seem to to be the
case when multiplying or adding whole numbers.